+++ config/tipoff.php
@@ -18,6 +18,8 @@ return [
 
     'model_class' => [
 
+        'address' => \Tipoff\Addresses\Models\Address::class,
+
         'block' => \Tipoff\Scheduler\Models\Block::class,
 
         'booking' => \Tipoff\Bookings\Models\Booking::class,
@@ -26,13 +28,19 @@ return [
 
         'cart_item' => \Tipoff\Checkout\Models\CartItem::class,
 
+        'city' => \Tipoff\Addresses\Models\City::class,
+
         'competitor' => \Tipoff\Reviews\Models\Competitor::class,
+
+        'company' => \Tipoff\Seo\Models\Company::class,
+
+        'country' => \Tipoff\Addresses\Models\Country::class,
 
         'contact' => \Tipoff\Forms\Models\Contact::class,
 
-        'customer' => \Tipoff\Addresses\Models\Customer::class, // Will be renamed later to address when new features added
-
         'discount' => \Tipoff\Discounts\Models\Discount::class,
+
+        'domain' => \Tipoff\Seo\Models\Domain::class,
 
         'fee' => \Tipoff\Fees\Models\Fee::class,
 
@@ -41,8 +49,12 @@ return [
         'flex_day' => \Tipoff\FlexScheduling\Models\FlexDay::class,
 
         'game' => \Tipoff\Scheduler\Models\Game::class,
-        
+
         'gmb_account' => \Tipoff\GoogleApi\Models\GmbAccount::class,
+
+        'gmb_detail' => \Tipoff\Locations\Models\GmbDetail::class,
+
+        'gmb_hour' =>  \Tipoff\Locations\Models\GmbHour::class,
 
         'image' => \DrewRoberts\Media\Models\Image::class,
 
@@ -51,6 +63,8 @@ return [
         'invoice' => \Tipoff\Invoices\Models\Invoice::class,
 
         'key' => \Tipoff\GoogleApi\Models\Key::class,
+
+        'keyword' => \Tipoff\Seo\Models\Keyword::class,
 
         'location' => \Tipoff\Locations\Models\Location::class,
 
@@ -68,19 +82,35 @@ return [
 
         'payment' => \Tipoff\Payments\Models\Payment::class,
 
+        'place' => \Tipoff\Seo\Models\Place::class,
+
+        'place_details' => \Tipoff\Seo\Models\PlaceDetails::class,
+
+        'place_hours' => \Tipoff\Seo\Models\PlaceHours::class,
+
         'post' => \DrewRoberts\Blog\Models\Post::class,
 
+        'profile_link' => \Tipoff\Seo\Models\ProfileLink::class,
+
         'rate' => \Tipoff\EscapeRoom\Models\Rate::class,
+
+        'ranking' => \Tipoff\Seo\Models\Ranking::class,
 
         'recurring_schedule' => \Tipoff\Scheduler\Models\RecurringSchedule::class,
 
         'refund' => \Tipoff\Refunds\Models\Refund::class,
+
+        'region' => \Tipoff\Addresses\Models\Region::class,
+
+        'result' => \Tipoff\Seo\Models\Result::class,
 
         'review' => \Tipoff\Reviews\Models\Review::class,
 
         'room' => \Tipoff\EscapeRoom\Models\Room::class,
 
         'schedule_eraser' => \Tipoff\Scheduler\Models\ScheduleEraser::class,
+
+        'search_volume' => \Tipoff\Seo\Models\SearchVolume::class,
 
         'series' => \DrewRoberts\Blog\Models\Series::class,
 
@@ -89,6 +119,8 @@ return [
         'slot' => \Tipoff\Scheduler\Models\Slot::class,
 
         'snapshot' => \Tipoff\Reviews\Models\Snapshot::class,
+
+        'state' => \Tipoff\Addresses\Models\State::class,
 
         'status' => \Tipoff\Statuses\Models\Status::class,
 
@@ -99,8 +131,8 @@ return [
         'tax' => \Tipoff\Taxes\Models\Tax::class,
 
         'theme' => \Tipoff\EscapeRoom\Models\Theme::class,
-        
-        'timezone' => \Tipoff\Locations\Models\Timezone::class,
+
+        'timezone' => \Tipoff\Addresses\Models\Timezone::class,
 
         'topic' => \DrewRoberts\Blog\Models\Topic::class,
 
@@ -112,9 +144,15 @@ return [
 
         'voucher_type' => \Tipoff\Vouchers\Models\VoucherType::class,
 
+        'webpage' => \Tipoff\Seo\Models\Webpage::class,
+
+        'zip' => \Tipoff\Addresses\Models\Zip::class,
+
     ],
 
     'nova_class' => [
+
+        'address' => \Tipoff\Addresses\Nova\Address::class,
 
         'block' => \Tipoff\Scheduler\Nova\Block::class,
 
@@ -124,13 +162,19 @@ return [
 
         'cart_item' => \Tipoff\Checkout\Nova\CartItem::class,
 
+        'city' => \Tipoff\Addresses\Nova\City::class,
+
+        'company' => \Tipoff\Seo\Nova\Company::class,
+
         'competitor' => \Tipoff\Reviews\Nova\Competitor::class,
 
         'contact' => \Tipoff\Forms\Nova\Contact::class,
 
-        'customer' => \Tipoff\Addresses\Nova\Customer::class, // Will be renamed later to address when new features added
+        'country' => \Tipoff\Addresses\Nova\Country::class,
 
         'discount' => \Tipoff\Discounts\Nova\Discount::class,
+
+        'domain' => \Tipoff\Seo\Nova\Domain::class,
 
         'fee' => \Tipoff\Fees\Nova\Fee::class,
 
@@ -139,7 +183,7 @@ return [
         'flex_day' => \Tipoff\FlexScheduling\Nova\FlexDay::class,
 
         'game' => \Tipoff\Scheduler\Nova\Game::class,
-        
+
         'gmb_account' => \Tipoff\GoogleApi\Nova\GmbAccount::class,
 
         'image' => \DrewRoberts\Media\Nova\Image::class,
@@ -149,6 +193,8 @@ return [
         'invoice' => \Tipoff\Invoices\Nova\Invoice::class,
 
         'key' => \Tipoff\GoogleApi\Nova\Key::class,
+
+        'keyword' => \Tipoff\Seo\Nova\Keyword::class,
 
         'location' => \Tipoff\Locations\Nova\Location::class,
 
@@ -166,7 +212,17 @@ return [
 
         'payment' => \Tipoff\Payments\Nova\Payment::class,
 
+        'place' => \Tipoff\Seo\Nova\Place::class,
+
+        'place_details' => \Tipoff\Seo\Nova\PlaceDetails::class,
+        
+        'place_hours' => \Tipoff\Seo\Nova\PlaceHours::class,
+
         'post' => \DrewRoberts\Blog\Nova\Post::class,
+
+        'profile_link' => \Tipoff\Seo\Nova\ProfileLink::class,
+
+        'ranking' => \Tipoff\Seo\Nova\Ranking::class,
 
         'rate' => \Tipoff\EscapeRoom\Nova\Rate::class,
 
@@ -174,11 +230,17 @@ return [
 
         'refund' => \Tipoff\Refunds\Nova\Refund::class,
 
+        'region' => \Tipoff\Addresses\Nova\Region::class,
+
+        'result' => \Tipoff\Seo\Nova\Result::class,
+
         'review' => \Tipoff\Reviews\Nova\Review::class,
 
         'room' => \Tipoff\EscapeRoom\Nova\Room::class,
 
         'schedule_eraser' => \Tipoff\Scheduler\Nova\ScheduleEraser::class,
+
+        'search_volume' => \Tipoff\Seo\Nova\SearchVolume::class,
 
         'series' => \DrewRoberts\Blog\Nova\Series::class,
 
@@ -187,6 +249,8 @@ return [
         'slot' => \Tipoff\Scheduler\Nova\Slot::class,
 
         'snapshot' => \Tipoff\Reviews\Nova\Snapshot::class,
+
+        'state' => \Tipoff\Addresses\Nova\State::class,
 
         'status' => \Tipoff\Statuses\Nova\Status::class,
 
@@ -197,8 +261,8 @@ return [
         'tax' => \Tipoff\Taxes\Nova\Tax::class,
 
         'theme' => \Tipoff\EscapeRoom\Nova\Theme::class,
-        
-        'timezone' => \Tipoff\Locations\Nova\Timezone::class,
+
+        'timezone' => \Tipoff\Addresses\Nova\Timezone::class,
 
         'topic' => \DrewRoberts\Blog\Nova\Topic::class,
 
@@ -209,6 +273,10 @@ return [
         'voucher' => \Tipoff\Vouchers\Nova\Voucher::class,
 
         'voucher_type' => \Tipoff\Vouchers\Nova\VoucherType::class,
+
+        'webpage' => \Tipoff\Seo\Nova\Webpage::class,
+
+        'zip' => \Tipoff\Addresses\Nova\Zip::class,
 
     ]